+++ src/compas_timber/connections/t_butt.py
@@ -56,6 +56,7 @@ class TButtJoint(ButtJoint):
             conical_tool=conical_tool,
             **kwargs,
         )
+        self.modify_cross = False
 
         self.fasteners = []
         if fastener: